--- Assets/RatHitbox.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ad2bde121925b4d428bc029d04303468
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
