+++ Assets/Scripts/NPC/RatHitbox.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class RatHitbox : MonoBehaviour
+{
+    [SerializeField] private float timeBetweenHits;
+    
+    void Start(){
+        
+    }
+    void Update(){
+        
+    }
+    private void OnTriggerEnter2D(Collider2D other) {
+        if(other.CompareTag("Player")){
+            other.transform.GetComponent<HP>().Damage(1,transform,0);
+            other.transform.GetComponent<hpSystem>().hpBarChange();
+        }
+    }
+}
