+++ Assets/Sprites/particles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b1849eb5e743ab9489c85502e197c852
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
