+++ Assets/Sprites/particles/Particle System Variant.prefab
@@ -1,0 +1,139 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &5846152672163598628
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 378596797122750374, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_Mesh
+      value: 
+      objectReference: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
+    - target: {fileID: 378596797122750374, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_SortMode
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 378596797122750374, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_RenderMode
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: lengthInSec
+      value: 3
+      objectReference: {fileID: 0}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.mode
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: InitialModule.size3D
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.sprites.Array.size
+      value: 9
+      objectReference: {fileID: 0}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: InitialModule.startSize.scalar
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: InitialModule.startSpeed.scalar
+      value: 1.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: EmissionModule.rateOverTime.scalar
+      value: 2
+      objectReference: {fileID: 0}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: InitialModule.startLifetime.scalar
+      value: 3
+      objectReference: {fileID: 0}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.sprites.Array.data[0].sprite
+      value: 
+      objectReference: {fileID: 27207027, guid: 9d08c5aa344324d41840635d1b1a711f, type: 3}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.sprites.Array.data[1].sprite
+      value: 
+      objectReference: {fileID: 14769583, guid: 9d08c5aa344324d41840635d1b1a711f, type: 3}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.sprites.Array.data[2].sprite
+      value: 
+      objectReference: {fileID: 823656388, guid: 9d08c5aa344324d41840635d1b1a711f, type: 3}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.sprites.Array.data[3].sprite
+      value: 
+      objectReference: {fileID: 1711655507, guid: 9d08c5aa344324d41840635d1b1a711f, type: 3}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.sprites.Array.data[4].sprite
+      value: 
+      objectReference: {fileID: 1561697940, guid: 9d08c5aa344324d41840635d1b1a711f, type: 3}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.sprites.Array.data[5].sprite
+      value: 
+      objectReference: {fileID: 1459651744, guid: 9d08c5aa344324d41840635d1b1a711f, type: 3}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.sprites.Array.data[6].sprite
+      value: 
+      objectReference: {fileID: -1744338242, guid: 9d08c5aa344324d41840635d1b1a711f, type: 3}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.sprites.Array.data[7].sprite
+      value: 
+      objectReference: {fileID: -1726168339, guid: 9d08c5aa344324d41840635d1b1a711f, type: 3}
+    - target: {fileID: 2594050264149038497, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: UVModule.sprites.Array.data[8].sprite
+      value: 
+      objectReference: {fileID: -717554323, guid: 9d08c5aa344324d41840635d1b1a711f, type: 3}
+    - target: {fileID: 5191576263387604830, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5191576263387604830, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -0.74
+      objectReference: {fileID: 0}
+    - target: {fileID: 5191576263387604830, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5191576263387604830, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5191576263387604830, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5191576263387604830, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5191576263387604830, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5191576263387604830, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: -90
+      objectReference: {fileID: 0}
+    - target: {fileID: 5191576263387604830, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5191576263387604830, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6165890716143034161, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
+      propertyPath: m_Name
+      value: Particle System Variant
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 48963d6e56d3bc94f83aaf3545d923ce, type: 3}
