+++ Assets/Sprites/particles/Particle System Variant.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a0517eae2ab3c434085c0f8a077f66e7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
