+++ Assets/Sprites/particles/Particle System.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 48963d6e56d3bc94f83aaf3545d923ce
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
